+++ Flow/test/utopia/flow/test/StringUtilsTest.scala
@@ -68,7 +68,7 @@ object StringUtilsTest extends App
 	
 	assert("XtestXX2".splitIterator("X").toVector == Vector("test", "2"))
 	
-	println("YV- 2716".filterWith(Regex.alphaNumeric))
+	println("YV- 2716".filterWith(Regex.letterOrDigit))
 	
 	println("Success!")
 }